feat(Mappings): consolidate storage/network mappings pages
[issue-112] marry network/storagem mapping pages
