--- src/app/Mappings/NetworkMappingsPage.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import { MappingType } from '@app/queries/types';
-import MappingsPage from './MappingsPage';
-
-const NetworkMappingsPage: React.FunctionComponent = () => (
-  <MappingsPage mappingType={MappingType.Network} key="network" />
-);
-
-export default NetworkMappingsPage;
