--- src/app/Mappings/StorageMappingsPage.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import { MappingType } from '@app/queries/types';
-import MappingsPage from './MappingsPage';
-
-const StorageMappingsPage: React.FunctionComponent = () => (
-  <MappingsPage mappingType={MappingType.Storage} key="storage" />
-);
-
-export default StorageMappingsPage;
