+++ src/app/Mappings/components/CreateMappingButton.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { Button, ButtonProps } from '@patternfly/react-core';
-import spacing from '@patternfly/react-styles/css/utilities/Spacing/spacing';
 import ConditionalTooltip from '@app/common/components/ConditionalTooltip';
 import { useHasSufficientProvidersQuery } from '@app/queries';
 
@@ -22,7 +21,7 @@ const CreateMappingButton: React.FunctionComponent<ICreateMappingButtonProps> = 
       isTooltipEnabled={!hasSufficientProviders}
       content="You must add at least one VMware provider and one OpenShift Virtualization provider in order to create a mapping."
     >
-      <div className={`${spacing.mtMd}`}>
+      <div>
         <Button onClick={onClick} isDisabled={!hasSufficientProviders} variant={variant}>
           {label}
         </Button>
