+++ src/app/utils/useDocumentTitle.ts
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
 // a custom hook for setting the page title
-export function useDocumentTitle(title: string) {
+export function useDocumentTitle(title: string): void {
   React.useEffect(() => {
     const originalTitle = document.title;
     document.title = title;
